forecasting/requirements.txt
# Prophet Forecasting Requirements
prophet==1.1.5
pandas==2.1.4
numpy==1.26.3
google-cloud-bigquery==3.14.1
db-dtypes==1.2.0
pyarrow==14.0.2
matplotlib==3.8.2
scikit-learn==1.3.2
